+++ asyncpraw/models/reddit/submission.py
@@ -495,48 +495,6 @@ class Submission(SubmissionListingMixin, UserContentMixin, FullnameMixin, Reddit
         """Return the class's kind."""
         return self._reddit.config.kinds["submission"]
 
-    async def comments(self) -> CommentForest:
-        """Provide an instance of :class:`.CommentForest`.
-
-        This attribute can be used, for example, to obtain a flat list of comments, with
-        any :class:`.MoreComments` removed:
-
-        .. code-block:: python
-
-            comments = await submission.comments()
-            await comments.replace_more(limit=0)
-            comment_list = await comments.list()
-            for comment in comment_list:
-                # do stuff with comment
-                ...
-
-        Sort order and comment limit can be set with the ``comment_sort`` and
-        ``comment_limit`` attributes before comments are fetched, including any call to
-        :meth:`.replace_more`:
-
-        .. code-block:: python
-
-            submission.comment_sort = "new"
-            comments = await submission.comments()
-            comment_list = await comments.list()
-            for comment in comment_list:
-                # do stuff with comment
-                ...
-
-        .. note::
-
-            The appropriate values for ``comment_sort`` include ``confidence``,
-            ``controversial``, ``new``, ``old``, ``q&a``, and ``top``
-
-        See :ref:`extracting_comments` for more on working with a
-        :class:`.CommentForest`.
-
-        """
-        # This assumes _comments is set so that _fetch is called when it's not.
-        if "_comments" not in self.__dict__:
-            await self._fetch()
-        return self._comments
-
     @cachedproperty
     def flair(self) -> SubmissionFlair:
         """Provide an instance of :class:`.SubmissionFlair`.
@@ -591,7 +549,7 @@ class Submission(SubmissionListingMixin, UserContentMixin, FullnameMixin, Reddit
 
         :param reddit: An instance of :class:`~.Reddit`.
         :param id: A reddit base36 submission ID, e.g., ``2gmzqe``.
-        :param url: A URL supported by :meth:`~asyncpraw.models.Submission.id_from_url`.
+        :param url: A URL supported by :meth:`~.id_from_url`.
 
         Either ``id`` or ``url`` can be provided, but not both.
 
@@ -611,6 +569,40 @@ class Submission(SubmissionListingMixin, UserContentMixin, FullnameMixin, Reddit
         super().__init__(reddit, _data=_data)
 
         self._comments_by_id = {}
+        self.comments = CommentForest(self)
+        """Provide an instance of :class:`.CommentForest`.
+
+        This attribute can be used, for example, to obtain a flat list of comments, with
+        any :class:`.MoreComments` removed:
+
+        .. code-block:: python
+
+            await submission.comments.replace_more(limit=0)
+            comments = submission.comments.list()
+
+        :raises: :py:class:`TypeError` if the submission is not fetched.
+
+        Sort order and comment limit must be set with the ``comment_sort`` and
+        ``comment_limit`` attributes before the submission and its comments are fetched,
+        including any call to :meth:`.replace_more`. The ``fetch`` argument will need to
+        set when initializing the :class:`.Submission` instance:
+
+        .. code-block:: python
+
+            submission = await reddit.submission("8dmv8z", fetch=False)
+            submission.comment_sort = "new"
+            await submission.load()
+            comments = submission.comments.list()
+
+        .. note::
+
+            The appropriate values for ``comment_sort`` include ``confidence``,
+            ``controversial``, ``new``, ``old``, ``q&a``, and ``top``
+
+        See :ref:`extracting_comments` for more on working with a
+        :class:`.CommentForest`.
+
+        """
 
     def __setattr__(self, attribute: str, value: Any):
         """Objectify author, subreddit, and poll data attributes."""
@@ -662,12 +654,11 @@ class Submission(SubmissionListingMixin, UserContentMixin, FullnameMixin, Reddit
         submission = type(self)(self._reddit, _data=submission_data)
         delattr(submission, "comment_limit")
         delattr(submission, "comment_sort")
-        submission._comments = CommentForest(self)
+        submission.comments = CommentForest(self)
 
         self.__dict__.update(submission.__dict__)
-        self._comments._update(comment_listing.children)
-
         self._fetched = True
+        self.comments._update(comment_listing.children)
 
     async def mark_visited(self):
         """Mark submission as visited.
@@ -720,7 +711,7 @@ class Submission(SubmissionListingMixin, UserContentMixin, FullnameMixin, Reddit
 
         .. code-block:: python
 
-            submission = await reddit.submission(id="5or86n")
+            submission = await reddit.submission(id="5or86n", fetch=False)
             await submission.unhide()
 
         .. seealso::