+++ CHANGES.rst
@@ -14,6 +14,12 @@ Unreleased
 - :meth:`.user_selectable` to get available subreddit link flairs.
 - Automatic RateLimit handling will support errors with millisecond resolution.
 
+**Deprecated**
+
+- Ability to use :class:`.CommentForest` as an asynchronous iterator.
+- :meth:`.CommentForest.list` no longer needs to be awaited.
+- :attr:`.Submission.comments` no longer needs to be awaited and is now a property.
+
 **Fixed**
 
 - Fixed return value type of methods returning a listing in :class:`.Subreddit` and its